# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.3.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /findings/{id}:
    get:
      summary: Fetch one finding with its six-step prose
      operationId: getFinding
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/repositories/{id}/findings:
    get:
      summary: Export findings for a repository as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `FindingExport` carrying the
        full set of `db.Finding` columns.
      operationId: exportRepositoryFindings
      security: []
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: format
          in: query
          description: Only `jsonl` is supported; other values get a 400.
          schema: { type: string, enum: [jsonl] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: JSONL stream of FindingExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
        '400':
          description: Unsupported format query value
        '403':
          description: Host header is not localhost
        '404':
          description: Repository not found
  /v1/findings:
    get:
      summary: Export findings across all repositories as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `FindingExport` carrying the
        full set of `db.Finding` columns.
      operationId: exportFindings
      security: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [jsonl] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: JSONL stream of FindingExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
        '400':
          description: Unsupported format query value
        '403':
          description: Host header is not localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    FindingSummary:
      type: object
      required: [id, scan_id, repository_id, severity, location]
      properties:
        id:            { type: integer, format: int64 }
        scan_id:       { type: integer, format: int64 }
        repository_id: { type: integer, format: int64 }
        finding_id:    { type: string, description: Short id within the scan report (F1, F2, ...). }
        sinks:         { type: string, description: Comma-joined sink ids. }
        title:         { type: string }
        severity:      { type: string, enum: [Critical, High, Medium, Low] }
        status:        { type: string }
        cwe:           { type: string }
        location:      { type: string }
        affected:      { type: string }
        reachability:  { type: string, enum: [reachable, harness_only, unclear] }
        quality_tier:  { type: string, enum: [high, low] }
        cve_id:        { type: string }
        cvss_vector:   { type: string }
        cvss_score:    { type: number }
        fix_version:   { type: string }
        fix_commit:    { type: string }
        resolution:    { type: string, enum: [fix, migrate, workaround, adopt, wontfix, ""] }
        assignee:      { type: string }
    FindingDetail:
      allOf:
        - $ref: '#/components/schemas/FindingSummary'
        - type: object
          description: Adds the six-step narrative and disclosure draft text.
          properties:
            trace:            { type: string }
            boundary:         { type: string }
            validation:       { type: string }
            prior_art:        { type: string }
            reach:            { type: string }
            rating:           { type: string }
            disclosure_draft: { type: string }
    FindingExport:
      description: |
        Full db.Finding row used by the unauthenticated JSONL export. Adds
        the dedup, denormalised, and timestamp columns that the bearer-only
        FindingDetail leaves out.
      allOf:
        - $ref: '#/components/schemas/FindingDetail'
        - type: object
          properties:
            commit:            { type: string }
            sub_path:          { type: string }
            fingerprint:       { type: string }
            last_seen_scan_id: { type: integer, format: int64 }
            last_seen_commit:  { type: string }
            seen_count:        { type: integer }
            created_at:        { type: string, format: date-time }
            updated_at:        { type: string, format: date-time }
...